Load Client data, generalize forms, Live web page, remove automation (#43)
* Moderate restructuring to account for new data

* Update Readme with more commands

* Alter structure to allow for dynamic forms

This creates a new structure so that the name of the model is passed dynamically, allowing us to reuse the same form in forms.py, as well as the same functions in views.py. The model, as well as important fields for the model that are required to be filled before creation, are stored as a tuple accessed by the models name, which is passed vie the url

* Update manage.py

* Re introduce netcat command for delay

* Data rework

* Create a generalized csv for loading data

This includes all the necessary fields from models.py

* Mount application and reduce automation

I think the automated migrations was causing some issues, so I'm removing them for now. This also mounts the project directory and data that needs to be loaded. Mounting the proj dir allows us to see any changes we make within the app right away.

* Add mysql-client to the django image so we can access sql directly from the Django container

* Create load_csv.py

This script provides a repeatable method for loading the csv data from our client. Can be modified as needed

* Altered models to reflect the data properly

* remove --noinput command

I believe this command was overly-supressing important warnings and errors

* Modify currchemicals.html to match new data

* Rework readme for readability

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Fixup the merge for add chemical changes

* Update README.md

* Update docker-compose.yml

* Update docker-compose.yml